Ensure that both branches of min() and max() resolve to the same type
This prepares the code to the c++ conversion

+++ src/char/char.c
@@ -1872,7 +1872,7 @@ static int char_delete_char_sql(int char_id)
 	SQL->GetData(inter->sql_handle, 9, &data, NULL);
 	elemental_id = atoi(data);
 
-	SQL->EscapeStringLen(inter->sql_handle, esc_name, name, min(len, NAME_LENGTH));
+	SQL->EscapeStringLen(inter->sql_handle, esc_name, name, min(len, (size_t)NAME_LENGTH));
 	SQL->FreeResult(inter->sql_handle);
 
 	//check for config char del condition [Lupus]
@@ -2029,19 +2029,19 @@ static int char_mmo_char_tobuf(uint8 *buffer, struct mmo_charstatus *p)
 
 	WBUFL(buf,0) = p->char_id;
 #if PACKETVER >= 20170830
-	WBUFQ(buf,4) = min(p->base_exp, INT64_MAX);
+	WBUFQ(buf,4) = min(p->base_exp, (uint64)INT64_MAX);
 	offset += 4;
 	buf = WBUFP(uint8 *, buffer, offset);
 #else
-	WBUFL(buf,4) = min((uint32)(p->base_exp), INT32_MAX);
+	WBUFL(buf,4) = min((uint32)(p->base_exp), (uint32)INT32_MAX); // FIXME: This should run on the larger type and be cast to uint32 after the clamping
 #endif
 	WBUFL(buf,8) = p->zeny;
 #if PACKETVER >= 20170830
-	WBUFQ(buf,12) = min(p->job_exp, INT64_MAX);
+	WBUFQ(buf,12) = min(p->job_exp, (uint64)INT64_MAX);
 	offset += 4;
 	buf = WBUFP(uint8 *, buffer, offset);
 #else
-	WBUFL(buf,12) = min((uint32)(p->job_exp), INT32_MAX);
+	WBUFL(buf,12) = min((uint32)(p->job_exp), (uint32)INT32_MAX); // FIXME: This should run on the larger type and be cast to uint32 after the clamping
 #endif
 	WBUFL(buf,16) = p->job_level;
 	WBUFL(buf,20) = 0; // probably opt1
@@ -2049,7 +2049,7 @@ static int char_mmo_char_tobuf(uint8 *buffer, struct mmo_charstatus *p)
 	WBUFL(buf,28) = (p->option &~ 0x40);
 	WBUFL(buf,32) = p->karma;
 	WBUFL(buf,36) = p->manner;
-	WBUFW(buf,40) = min(p->status_point, INT16_MAX);
+	WBUFW(buf,40) = min(p->status_point, (int)INT16_MAX);
 #if PACKETVER_MAIN_NUM >= 20201007 || PACKETVER_RE_NUM >= 20211103 || PACKETVER_ZERO_NUM >= 20221024
 	WBUFQ(buf, 42) = p->hp;
 	WBUFQ(buf, 50) = p->max_hp;
@@ -2061,8 +2061,8 @@ static int char_mmo_char_tobuf(uint8 *buffer, struct mmo_charstatus *p)
 	offset+=4;
 	buf = WBUFP(uint8 *, buffer,offset);
 #else
-	WBUFW(buf,42) = min(p->hp, INT16_MAX);
-	WBUFW(buf,44) = min(p->max_hp, INT16_MAX);
+	WBUFW(buf,42) = min(p->hp, (int)INT16_MAX);
+	WBUFW(buf,44) = min(p->max_hp, (int)INT16_MAX);
 #endif
 #if PACKETVER_MAIN_NUM >= 20201007 || PACKETVER_RE_NUM >= 20211103 || PACKETVER_ZERO_NUM >= 20221024
 	WBUFQ(buf, 46) = p->sp;
@@ -2070,8 +2070,8 @@ static int char_mmo_char_tobuf(uint8 *buffer, struct mmo_charstatus *p)
 	offset += 12;
 	buf = WBUFP(uint8 *, buffer, offset);
 #else  // PACKETVER_MAIN_NUM >= 20201007 || PACKETVER_RE_NUM >= 20211103 || PACKETVER_ZERO_NUM >= 20221024
-	WBUFW(buf, 46) = min(p->sp, INT16_MAX);
-	WBUFW(buf, 48) = min(p->max_sp, INT16_MAX);
+	WBUFW(buf, 46) = min(p->sp, (int)INT16_MAX);
+	WBUFW(buf, 48) = min(p->max_sp, (int)INT16_MAX);
 #endif  // PACKETVER_MAIN_NUM >= 20201007 || PACKETVER_RE_NUM >= 20211103 || PACKETVER_ZERO_NUM >= 20221024
 	WBUFW(buf,50) = DEFAULT_WALK_SPEED; // p->speed;
 	WBUFW(buf,52) = p->class_;
@@ -2087,7 +2087,7 @@ static int char_mmo_char_tobuf(uint8 *buffer, struct mmo_charstatus *p)
 	WBUFW(buf,56) = (p->option&(OPTION_RIDING|OPTION_DRAGON|OPTION_WUG|OPTION_WUGRIDER|OPTION_MADOGEAR|OPTION_HANBOK)) ? 0 : p->look.weapon;
 
 	WBUFW(buf,58) = p->base_level;
-	WBUFW(buf,60) = min(p->skill_point, INT16_MAX);
+	WBUFW(buf,60) = min(p->skill_point, (int)INT16_MAX);
 	WBUFW(buf,62) = p->look.head_bottom;
 	WBUFW(buf,64) = p->look.shield;
 	WBUFW(buf,66) = p->look.head_top;
@@ -2095,12 +2095,12 @@ static int char_mmo_char_tobuf(uint8 *buffer, struct mmo_charstatus *p)
 	WBUFW(buf,70) = p->hair_color;
 	WBUFW(buf,72) = p->clothes_color;
 	memcpy(WBUFP(char *, buf,74), p->name, NAME_LENGTH);
-	WBUFB(buf,98) = min(p->str, UINT8_MAX);
-	WBUFB(buf,99) = min(p->agi, UINT8_MAX);
-	WBUFB(buf,100) = min(p->vit, UINT8_MAX);
-	WBUFB(buf,101) = min(p->int_, UINT8_MAX);
-	WBUFB(buf,102) = min(p->dex, UINT8_MAX);
-	WBUFB(buf,103) = min(p->luk, UINT8_MAX);
+	WBUFB(buf,98) = min(p->str, (short)UINT8_MAX);
+	WBUFB(buf,99) = min(p->agi, (short)UINT8_MAX);
+	WBUFB(buf,100) = min(p->vit, (short)UINT8_MAX);
+	WBUFB(buf,101) = min(p->int_, (short)UINT8_MAX);
+	WBUFB(buf,102) = min(p->dex, (short)UINT8_MAX);
+	WBUFB(buf,103) = min(p->luk, (short)UINT8_MAX);
 	WBUFW(buf,104) = p->slot;
 #if PACKETVER >= 20061023
 	WBUFW(buf,106) = ( p->rename > 0 ) ? 0 : 1;
@@ -2998,7 +2998,7 @@ static void char_read_fame_list(void)
 		smith_fame_list[i].fame = atoi(data);
 		// name
 		SQL->GetData(inter->sql_handle, 2, &data, &len);
-		memcpy(smith_fame_list[i].name, data, min(len, NAME_LENGTH));
+		memcpy(smith_fame_list[i].name, data, min(len, (size_t)NAME_LENGTH));
 	}
 	// Build Alchemist ranking list
 	if( SQL_ERROR == SQL->Query(inter->sql_handle, "SELECT `char_id`,`fame`,`name` FROM `%s` WHERE `fame`>0 AND (`class`='%d' OR `class`='%d' OR `class`='%d' OR `class`='%d' OR `class`='%d' OR `class`='%d') ORDER BY `fame` DESC LIMIT 0,%d", char_db, JOB_ALCHEMIST, JOB_CREATOR, JOB_BABY_ALCHEMIST, JOB_GENETIC, JOB_GENETIC_T, JOB_BABY_GENETIC, fame_list_size_chemist) )
@@ -3013,7 +3013,7 @@ static void char_read_fame_list(void)
 		chemist_fame_list[i].fame = atoi(data);
 		// name
 		SQL->GetData(inter->sql_handle, 2, &data, &len);
-		memcpy(chemist_fame_list[i].name, data, min(len, NAME_LENGTH));
+		memcpy(chemist_fame_list[i].name, data, min(len, (size_t)NAME_LENGTH));
 	}
 	// Build Taekwon ranking list
 	if( SQL_ERROR == SQL->Query(inter->sql_handle, "SELECT `char_id`,`fame`,`name` FROM `%s` WHERE `fame`>0 AND (`class` in('%d', '%d')) ORDER BY `fame` DESC LIMIT 0,%d", char_db, JOB_TAEKWON, JOB_BABY_TAEKWON, fame_list_size_taekwon) )
@@ -3028,7 +3028,7 @@ static void char_read_fame_list(void)
 		taekwon_fame_list[i].fame = atoi(data);
 		// name
 		SQL->GetData(inter->sql_handle, 2, &data, &len);
-		memcpy(taekwon_fame_list[i].name, data, min(len, NAME_LENGTH));
+		memcpy(taekwon_fame_list[i].name, data, min(len, (size_t)NAME_LENGTH));
 	}
 	SQL->FreeResult(inter->sql_handle);
 }
